requirements.txt
numpy>=1.22.0
opencv-python>=4.0.0.0
Pillow>=9.0.0
scikit-image>=0.18.0
tensorflow>=2.0.0
tqdm>=4.62